fix(cli): add missing example file to package, improve behavior when proxied
The CLI's help action tries to print an example configuration file, but that file wasn't getting
included in the NPM package, so it would just error out. That file has been added to the includes in
package.json.  Additionally, the js and css asset references for the CLI embedded application have
been moved to relative paths so that the app still works correctly if proxied under a non-root path
on a domain.