ci(versioning): fix tags not getting pushed back to origin
We're using `--follow-tags` when pushing to get release tags back to the origin, but it turns out
that `--follow-tags` only pushes annotated tags, which we weren't creating due to our monorepo
changes. This switches our tag command to create an annotated tag with a message, identical to what
was happening with the previous tooling.