feat(web-workers): Add web-worker support
This adds new `WorkerPool` and `WorkerClient` classes which are analagous
with the `frame-router` and `Client` classes, respectively. The client
classes have been refactored to share a common base implementation. The
`WorkerPool` class uses a base `HostManager` class, which will be
incorporated into the host side of iframe handling in a future update.